+++ CHANGES.md
@@ -13,6 +13,7 @@ Warning: this release introduces some changes which may break backwards compatib
 * [#77] Added: `MoreCollectors.filtering()/mapping()/flatMapping()` with default downstream Collector `toList()`.
 * [#79] Added: `iterate(seed, predicate, op)` for all stream types.
 * [#83] Changed: `StreamEx.of(Collection)`, `of(Iterator)`, etc. now use `? extends T` type instead of just `T`.
+* [#86] Added: `peekFirst()`, `peekLast()` for all stream types.
 
 
 ### 0.5.5